1.处理panic导致程序退出的问题 2. 优化services中.go
