--- services/delete_job.go
@@ -1,16 +1,0 @@
-package services
-
-import (
-	"context"
-	"fmt"
-	"pixiu/variable"
-)
-
-func DeleteJob() {
-	ctx := context.Background()
-	DeleteJob, err := Jenkins.DeleteJob(ctx, variable.DeleteJobName)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("job 已经删除%v", DeleteJob)
-}
